Crate terminal

A helper library for terminal output.

This library is used by Spin to print out messages in an appropriate format that is easy for users to read. This is not meant as a general purpose library.

Modules

colors

Macros

ceprint

ceprintln

cprint

einfo

error

step

warn

Structs

ColorText

A wrapper around a standard stream lock that resets the color on drop